+++ app/src/main/java/ru/practicum/android/diploma/vacancy/filter/ui/FilterViewModel.kt
@@ -7,12 +7,10 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import ru.practicum.android.diploma.vacancy.filter.domain.FilterSettingsInteractor
-import ru.practicum.android.diploma.vacancy.filter.domain.api.IndustryFilterInteractor
 import ru.practicum.android.diploma.vacancy.filter.domain.model.FilterIndustryValue
 import ru.practicum.android.diploma.vacancy.filter.domain.model.FilterSettings
 
 class FilterViewModel(
-    private val interactor: IndustryFilterInteractor,
     private val filterSettingsInteractor: FilterSettingsInteractor
 
 ) : ViewModel() {
@@ -26,23 +24,6 @@ class FilterViewModel(
     private val _filterSettings = MutableStateFlow<FilterSettings?>(null)
     val filterSettings: StateFlow<FilterSettings?> = _filterSettings
 
-    init {
-        loadSelectedIndustry()
-    }
-
-    // Загрузка сохраненной отрасли
-    private fun loadSelectedIndustry() {
-        _selectedIndustry.value = interactor.getSelectedIndustry()
-    }
-
-    // Сохранение выбранной отрасли
-    fun saveIndustry(industry: FilterIndustryValue) {
-        interactor.saveSelectedIndustry(industry)
-        _selectedIndustry.value = industry
-    }
-    fun setOnlyWithSalary(value: Boolean) {
-        _onlyWithSalary.value = value
-    }
 
     // Получить настройки фильтров
     fun loadFilterSettings() {
